break out classes into files, clean up code
