+++ Domain/Domain.csproj
@@ -139,6 +139,8 @@
     <Compile Include="IETagChecker.cs" />
     <Compile Include="IEventMigratingAggregate.cs" />
     <Compile Include="IStore{T}.cs" />
+    <Compile Include="Snapshots\IApplySnapshot{T}.cs" />
+    <Compile Include="Snapshots\JsonSnapshot.cs" />
     <Compile Include="Snapshots\JsonSnapshotter{T}.cs" />
     <Compile Include="Scheduling\CommandDeduplicated.cs" />
     <Compile Include="Scheduling\CommandDelivered.cs" />
